% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental.R
\name{elemental}
\alias{elemental}
\title{elemental}
\usage{
elemental(X, Y)
}
\arguments{
\item{X}{Name of stimulus vector (as string)}

\item{Y}{Name of stimulus vector (as string)}
}
\value{
Generalization value
}
\description{
Generalization function for elemental models.
}
\details{
elemental() computes generalization according the inner product
rule appropriate for elemental models of associative learning.
}
\author{
Stefano Ghirlanda
}
