% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{experiment}
\usage{
experiment(model, design, group.size)
}
\arguments{
\item{model}{An alsim model}

\item{design}{An experimental design}

\item{group.size}{Hown many subjects to run in each group}
}
\value{
An updated alsim model object
}
\description{
Run an experiment in alsim
}
\details{
An experiment is a list of groups, a group is a list of phases. For what a phase is see the documentation to phase.
}
\author{
Stefano Ghirlanda
}
