% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newModel.R
\name{newModel}
\alias{newModel}
\title{newModel}
\usage{
newModel(genFunction, learnRate)
}
\arguments{
\item{genFunction}{An alsim generalization function}

\item{learnRate}{A learning rate}
}
\value{
An alsim model
}
\description{
Create a new alsim model
}
\details{
newModel creates a model given a generalization function and a learning rate
}
\author{
Stefano Ghirlanda
}
