% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearce.R
\name{pearce}
\alias{pearce}
\title{pearce}
\usage{
pearce(X, Y)
}
\arguments{
\item{X}{Name of stimulus vector (as string)}

\item{Y}{Name of stimulus vector (as string)}
}
\value{
Generalization value
}
\description{
Generalization function for Pearce's "configural" model.
}
\details{
rem() computes generalization according to Pearce's (1987,1994)
"configural" model of associative learning.  model.
}
\author{
Stefano Ghirlanda
}
