% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{phase}
\alias{phase}
\title{phase}
\usage{
phase(model, trials)
}
\arguments{
\item{model}{An alsim model object}

\item{trials}{A list of trial specifications (see details)}
}
\value{
An updated alsim model object
}
\description{
Run multiple trials with multiple stimuli
}
\details{
phase runs multiple learning trials with multiple stimuli. The
trials argument is a list whose names are stimulus names, with
each element a numeric vector giving the amount of trials and the
reinforcement. Thus list( A=c(50,1) ) runs a phase with 50
presentations of A and reinforcement of 1. When multiple stimuli
are used, the sequence of trials is randomized.
}
\author{
Stefano Ghirlanda
}
