% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{rem}
\alias{rem}
\title{rem}
\usage{
rem(X, Y)
}
\arguments{
\item{X}{Name of stimulus vector (as string)}

\item{Y}{Name of stimulus vector (as string)}
}
\value{
Generalization value
}
\description{
Generalization function for the "replaced elements" model.
}
\details{
rem() computes generalization based on the "replaced elements"
model.
}
\author{
Stefano Ghirlanda
}
