% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{remify}
\alias{remify}
\title{remify}
\usage{
remify(X)
}
\arguments{
\item{X}{Name of stimulus vector (as string)}
}
\value{
Generalization value
}
\description{
Modify stimulus vectors according to the "replaced elements" model.
}
\details{
remify(X) modifies stimulus vector X according to the
"replaced elements" model, so that the elemental() generalization
function can be used on remify'ed vectors to simulate the replaced
elements model.
}
\author{
Stefano Ghirlanda
}
