% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{trial}
\alias{trial}
\title{trial}
\usage{
trial(model, X, lambda)
}
\arguments{
\item{model}{An alsim model}

\item{X}{A stimulus name (as string)}

\item{lambda}{Desired associative strength}
}
\value{
Updated model
}
\description{
Perform a single learning trial
}
\details{
Performs a single learning trial on a model, given a stimulus X
and desired associative strength lambda.
}
\author{
Stefano Ghirlanda
}
