% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V.R
\name{V}
\alias{V}
\title{V}
\usage{
V(model, X, trial = NULL, aggr = TRUE)
}
\arguments{
\item{model}{An alsim model or list of models}

\item{X}{A stimulus name (string)}

\item{trial}{Trial or trials for which associative strength should
be calculated. If NULL, use the last available trial.}

\item{aggr}{If TRUE, average results for all models given in input
and calculate standard deviation. If FALSE, return results for
all models.}
}
\value{
If aggr==TRUE, a 2-dim matrix with rows corresponding to
  average and standard deviation, and columns corresponding to
  trials. If aggr==FALSE, a matrix with one row per model, and
  columns corresponding to trials.
}
\description{
Calculate associative strength
}
\details{
V calculates the associative strengths of stimuli for a given
model
}
\author{
Stefano Ghirlanda
}
