% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vplot.R
\name{Vplot}
\alias{Vplot}
\title{Vplot}
\usage{
Vplot(model, stimuli, trials, jitter = 0)
}
\arguments{
\item{model}{An alsim model}

\item{stimuli}{Stimuli whose associative strength should be
plotted.}

\item{trials}{Vector of trials for which associative strengths
should be plotted.}

\item{jitter}{Add vertical jitter. Sometimes useful to avoid
overplotting}
}
\value{

}
\description{
Plot associative strengths for an alsim model
}
\details{
MISSING
}
\author{

}
